moving pure functions to util
